refactor: renamed OrderInfo.jsx to .tsx
